LOGITECH HID VID

+++ Aplicacao Java/src/Logitech/HeadSet.java
@@ -7,7 +7,7 @@ public class HeadSet extends ILogitech implements IHeadSet {
     private final static int[] zonasRGB = {0, 1};
 
     public HeadSet(String nome, String ID, java.awt.Color cor) {
-        super(nome, ID, cor, zonasRGB, 8);
+        super(nome, ID, cor, zonasRGB,(short) 8);
     }
 
 }
