+++ Aplicacao Java/src/Logitech/Keyboard.java
@@ -7,7 +7,7 @@ public final class Keyboard extends ILogitech implements IKeyboard {
     private final static int[] zonasRGB = {0, 1, 2, 3, 4};
 
     public Keyboard(String nome, String ID, java.awt.Color cor) {
-        super(nome, ID, cor, zonasRGB, 0);
+        super(nome, ID, cor, zonasRGB, (short)0);
     }
 
     @Override
