+++ Aplicacao Java/src/Logitech/Speaker.java
@@ -7,7 +7,7 @@ public final class Speaker extends ILogitech implements ISpeaker {
     private final static int[] zonasRGB = {0, 1, 2, 3};
 
     public Speaker(String nome, String ID, java.awt.Color cor) {
-        super(nome, ID, cor, zonasRGB, 14);
+        super(nome, ID, cor, zonasRGB,(short) 14);
     }
 
 }
